rename developper -> developer departement -> department
