+++ com.github.avdyk.osgi.caching.test-api/src/main/java/com/github/avdyk/osgi/caching/test/api/DeveloperService.java
@@ -1,0 +1,8 @@
+package com.github.avdyk.osgi.caching.test.api;
+
+import java.util.List;
+
+public interface DeveloperService {
+
+    List<String> getDevelopers(String service);
+}
