--- com.github.avdyk.osgi.caching.test-api/src/main/java/com/github/avdyk/osgi/caching/test/api/DevelopperService.java
@@ -1,8 +1,0 @@
-package com.github.avdyk.osgi.caching.test.api;
-
-import java.util.List;
-
-public interface DevelopperService {
-
-    List<String> getDeveloppers(String service);
-}
